Update Belgian RGOs
Move iron and coal provinces around to better represent history, results in a bit more iron and less coal. Add sulfur again in Charleroi to represent the local phosphate mining (Carrières de la Malogne). Add cotton in Bruges to represent cultivation of flax (as in Alençon, France) and to give the cloth industry a better chance of success.

379: coal --> wool (no coal here, some local textile industry)
387: coal --> grain (no coal here, grain moved from 388)
388: grain --> cotton (represents flax, cfr Alençon)
389: forest --> farmlands (no forest here)
390: farmlands --> plains (relatively poor agricultural ground)
391: coal --> grain (poor farming region, coal mines opened 1901)
392: farmlands --> mining_district (rich mining / industrial region)
395: iron --> coal (both present, compensates for other changes)
396: coal --> timber (area later given to Belgium specifically to compensate for looted timber in WW1)
397: coal --> iron (No coal here, Lorraine minette)
398: coal --> iron (No coal here, Lorraine minette, Wellin deposit)