Modified 'Alternative' Treaty of London
Belgium now starts with cores on provinces south of the rivers, where the majority of pops are catholic. The 'Let us demand concessions' option of the new alternate Treaty of London (after a Dutch surrender) now allows taking the Zeeland state instead of 2 random colonies, resulting in a Belgium / Netherlands split based on majority religion (catholic Belgium and protestant Netherlands).